Projects/News_Reporter_AI_Agent/requirements.txt
crewai
langchain_google_genai
load_dotenv
streamlit
crewai_tools